Refactor project to patternize
