+++ .gitignore
@@ -25,6 +25,7 @@ pkg/cm-beetle
 pkg/cm-beetle-arm
 pkg/cli/tbctl
 pkg/cli/cli
+cmd/cm-beetle/cm-beetle
 
 # AutoGenerated DB data directory
 meta_db/
