+++ pkg/api/rest/common/utility.go
